+++ provider/reprovider.go
@@ -411,6 +411,12 @@ func parseTime(b []byte) (time.Time, error) {
 	return time.Unix(0, tns), nil
 }
 
+// Clear removes all entries from the provide queue. Returns the number of CIDs
+// removed from the queue.
+func (s *reprovider) Clear() int {
+	return s.q.Clear()
+}
+
 func (s *reprovider) Close() error {
 	s.close()
 	err := s.q.Close()